Stylesheet implementation and styling website, adjustments
I have added a stylesheet that is referenced within the HTML on some of the pages. This adds white borders around individual objects and also colours the text to white, as the background itself is black. I've also centred the text to the center, so that it is all organised efficiently.

I have used flex boxes in order to position individual objects, making it so that the objects are responsive with the browser being used.