Updates:
In the process of removing white-spaces. Added a conditional check on the forum that checks whether or not the user is currently authenticated and displays to them whether they are not. I have also commented the code better, so that anyone reading it can understand better what the different mechanisms do, by referencing them in the comments.

I've also added a levels.php which is a prime example of interactivity with the SteamAPI for retrieving the level of the user. I am going to be providing users a link to be able to find out their own SteamId to avoid confusion.