Done it!
Upon creating a post the user will be redirected to the page of the post that they have created. In the SQL. statement,  I ordered the list and and added a limitation of 1 so that it would only return the newly created post from the database. That means that when the user clicks create posts and enters text into both input fields and submits that information, they are directed to the page with that content. I've done this, so that the user is easily able to see their page upon creating it and doesn't have to scroll down to do so.