Steam level calculator updates:
Validation for inputs have been implemented. This is to prevent users from inputting malicious code which could be potentially harmful to the database. I've also made it so that upon the users signup, the steam community id they have input is inserted into the database for later use. I am going to be implementing a feature which provides them with the ability to change this, assuming that they do change their own username on Steam.

I've added a short summary explaining how to use the calculator too, for anyone who may be confused when it comes to understanding it's purpose and how it may be used.